requirements.txt
langchain
langchain-core

langchain-community 
wikipedia

langchain-openai
openai

langchain-anthropic

langchain-google-genai
google-generativeai

langchain-huggingface
transformers
huggingface-hub

pydantic

python-dotenv
grandalf
streamlit

langchain_postgres
langchain-chroma
langchain-classic